ml_model/requirements.txt
nltk
torch
pymystem3
transformers
sentencepiece
sklearn
pandas
grpcio
google
protobuf
dynaconf